add authentication functionality and api service.
apologies for the enormous amount of changes. This commit includes the integration for our api 'ApiService'. It also includes an 'AuthenticationService' which alllows educators to register/login to the frontend. The login/registration mechanism stores a session token on the browsers local storage and is used to validate subsequent requests that require authorization. This includes api requests needed to write to the backend storage.